Correccion de atributo cantidad_minima, acepta decimales

+++ acqua-backend/database/migrations/2023_06_10_003335_create_servicios_table.php
@@ -18,7 +18,7 @@ class CreateServiciosTable extends Migration
             $table->string('clave_servicio')->unique();
             $table->string('nombre_servicio');
             $table->decimal('importe', 8, 2);
-            $table->integer('cantidad_minima');
+            $table->decimal('cantidad_minima', 8, 2);
             $table->timestamps();
             $table->foreign('catalogo_id')->references('id')->on('catalogos')->onDelete('cascade');
         });
